# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "BTEsse3"
  description: This API provide a service to handle exams from the perspective of the student and the teacher
  license:
    name: GPL3
    url: https://github.com/caramellaio/BTEsse3/blob/master/LICENSE
basePath: /
schemes:
- "http"
- "https"
consumes:
- application/json
produces:
- application/json
paths:
...
  /exams:
    get:
      tags:
      - "exam"
      summary: "Return all the exams"
      description: ""
      operationId: "getExams"
      responses: 
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "Exam not found"
    post:
      tags:
      - "exam"
      summary: "Create an exam"
      description: ""
      operationId: "createExam"
      parameters:
      - in: "body"
        name: "body"
        description: "Created exam object"
        required: true
        schema:
          $ref: "#/definitions/Exam"
      responses:
        200:
          description: "OK"
        400:
          description: "Exam creation failed"
  /exams/idUsers/{idUser}:
    get:
      tags:
      - "exam"
      summary: "Return the exams created by user"
      description: ""
      operationId: "getExamsbyUserId"
      parameters:
      - name: "idUser"
        in: "path"
        required: true
        description: "Return the exams created by user"
        type: "integer"
      responses: 
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "Exam not found"
  /exams/{idExam}:
    get:
      tags:
      - "exam"
      summary: "Return the exam by id"
      description: ""
      operationId: "getExamById"
      parameters:
      - name: "idExam"
        in: "path"
        required: true
        description: "Return the exam by id"
        type: "integer"
      responses: 
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "Exam not found"
    put:
      tags:
      - "exam"
      summary: "Update the exam"
      description: ""
      operationId: "updateExam"
      parameters:
      - name: "idExam"
        in: "path"
        description: "The id that needs to be deleted"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Created exam object"
        required: true
        schema:
          $ref: "#/definitions/Exam"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Exam"
        400:
          description: "Update failed"
    delete:
      tags:
      - "exam"
      summary: "Delete exam"
      description: ""
      operationId: "deleteExam"
      parameters:
      - name: "idExam"
        in: "path"
        description: "The id that needs to be deleted"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
        404:
          description: "Exam not found"
...
  /submissions:
    get:
      tags:
      - "submission"
      summary: "Get the submissions"
      description: ""
      operationId: "getSubmittedExam"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Submission"
        404:
          description: "Submissions not found"
    post:
      tags:
      - "submission"
      summary: "Submit exam"
      description: ""
      operationId: "submitExam"
      parameters:
      - in: "body"
        name: "body"
        description: "Submitted exam object"
        required: true
        schema:
          $ref: "#/definitions/Submission"
      responses:
        200:
          description: "OK"
        400:
          description: "Submission not well formed"
definitions:
...
  Task:
    type: "object"
    required:
    - "id"
    - "type"
    - "request"
    - "response"
    properties:
      id:
        type: "integer"
        format: "int64"
      type:
        type: "string"
      request:
        type: "string"
      response:
        type: "string"
      reviews:
        type: "array"
        xml:
          name: "tasks"
          wrapped: true
        items:
          $ref: "#/definitions/Review"
    xml:
      name: "Task"
  Exam:
    type: "object"
    required:
    - "nameExam"
    - "idExam"
    - "valutation"
    - "tasks"
    properties:
      nameExam:
        type: "string"
      idExam:
        type: "integer"
        format: "int64"
      idUser:
        type: "integer"
        format: "int64"
      valutation:
        type: "integer"
        format: "int64"
      tasks:
        type: "array"
        xml:
          name: "exams"
          wrapped: true
        items:
          $ref: "#/definitions/Task"
    xml:
      name: "TaskGroup"
...
  Submission:
    type: "object"
    required:
    - "id"
    - "idUser"
    - "exam"
    properties:
      id:
        type: "integer"
        format: "int64"
      idUser:
        type: "integer"
        format: "int64"
      exam:
        $ref: "#/definitions/Exam"
    xml:
      name: "Submission"
  Review:
    type: "object"
    required:
    - "idReview"
    - "idUser"
    - "idTask"
    - "message"
    properties:
      idReview:
        type: "integer"
        format: "int64"
      idUser:
        type: "integer"
      idTask:
        type: "integer"
      message:
        type: "string"
      mark:
        type: "integer"
    xml:
      name: "Review"